Move the default "Auto" method out into it's own class
